SFDNet Model/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.64.0
tensorboard>=2.10.0
lpips>=0.1.4
einops>=0.6.0
scikit-image>=0.19.0